+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+
+## v0.4.0
+
+Added
+
+- New Class: `InoreaderConfigManager` for config management
+
+Changed
+
+- Use OAuth2.0 authentication instead of user authentication with password
+- Optimized code of `InoreaderClient`
+- Optimized results of commands
+
 ## v0.3.0
 
 Added
@@ -35,7 +48,7 @@ Added
   - `InoreaderClient.mark_as_read`
   - `InoreaderClient.mark_as_starred`
   - `InoreaderClient.mark_as_liked`
-  - `InoreaderClient.boradcast`
+  - `InoreaderClient.broadcast`
 
 - New command `filter`